% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getHastagUserLinks}
\alias{getHastagUserLinks}
\title{Function to get user interaction links from tweets}
\usage{
getHastagUserLinks(tweets.df, exlclude_hashtags = FALSE)
}
\arguments{
\item{tweets.df}{data.frame of tweet messages and users}

\item{exlclude_hashtags}{default FALSE, whether to include hashtags as nodes in network}
}
\value{
link.df data.frame of links between users
}
\description{
Function to get user interaction links from tweets
}
