% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getLastTweet}
\alias{getLastTweet}
\title{Function to get last tweet of user}
\usage{
getLastTweet(userInfo)
}
\arguments{
\item{userInfo}{user object of the user for whose last tweet needs to be fetched}
}
\value{
time when user was last active
}
\description{
Function to get last tweet of user
}

