% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getNetworkLinksDF}
\alias{getNetworkLinksDF}
\title{Function to get network links with node ids from nodes dataframe and links dataframe}
\usage{
getNetworkLinksDF(links.df, network.nodes.df)
}
\arguments{
\item{links.df}{data.frame of links between users}

\item{network.nodes.df}{data.frame of nodes with ids}
}
\value{
network.links.df dataframe of links to plot network
}
\description{
Function to get network links with node ids from nodes dataframe and links dataframe
}
