% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getNetworkNodesDF}
\alias{getNetworkNodesDF}
\title{Function to get nodes data.frame for network from tweets and links dataframes}
\usage{
getNetworkNodesDF(tweets.df, links.df)
}
\arguments{
\item{tweets.df}{data.frame of tweet messages and users}

\item{links.df}{data.frame of links between users}
}
\value{
network.nodes.df dataframe of node in the network
}
\description{
Function to get nodes data.frame for network from tweets and links dataframes
}
