% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getTweetsDataFrame}
\alias{getTweetsDataFrame}
\title{Function to get tweets as clean data.frame from a tweets object}
\usage{
getTweetsDataFrame(tweets)
}
\arguments{
\item{tweets}{collection object returned by twitteR::search}
}
\value{
tweets.df a data.frame with tweets details
}
\description{
Function to get tweets as clean data.frame from a tweets object
}
