% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getTweetsInteractionNetwork}
\alias{getTweetsInteractionNetwork}
\title{Function to get interaction network from tweets datafrae}
\usage{
getTweetsInteractionNetwork(tweets.df, exlclude_hashtags = FALSE)
}
\arguments{
\item{tweets.df}{data.frame of tweet messages and users}

\item{exlclude_hashtags}{default FALSE, whether to include hashtags as nodes in network}
}
\value{
networkd.df dataframe of interaction network
}
\description{
Function to get interaction network from tweets datafrae
}
