% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserData}
\alias{getUserData}
\title{Function to gather data for user handles and save extracted data into a csv file}
\usage{
getUserData(userFile)
}
\arguments{
\item{userFile}{file path of the text file containing user handles}
}
\description{
Function to gather data for user handles and save extracted data into a csv file
}
