% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserFollowerCount}
\alias{getUserFollowerCount}
\title{Function to get total number of followers for a user}
\usage{
getUserFollowerCount(userInfo)
}
\arguments{
\item{userInfo}{user object of the user for whom  count need to be extracted}
}
\value{
count of followers of this user
}
\description{
Function to get total number of followers for a user
}

