% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserFollowers}
\alias{getUserFollowers}
\title{Function to get followers for a given user twitter handle}
\usage{
getUserFollowers(userInfo)
}
\arguments{
\item{userInfo}{user object of the user for whom followers need to be extracted}
}
\value{
vector of first degree followers for the user
}
\description{
Function to get followers for a given user twitter handle
}

