% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserInfo}
\alias{getUserInfo}
\title{#########################################################################################
basic api for twitter data exploration
#########################################################################################
Function to get user Info for a given user twitter handle}
\usage{
getUserInfo(username)
}
\arguments{
\item{username}{the twitter handle of user whose followers need to be extracted.}
}
\value{
userInfo object with information about user
}
\description{
#########################################################################################
basic api for twitter data exploration
#########################################################################################
Function to get user Info for a given user twitter handle
}

