% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserLinksFromTweets}
\alias{getUserLinksFromTweets}
\title{Function to get user interaction links from tweets}
\usage{
getUserLinksFromTweets(tweets.df)
}
\arguments{
\item{tweets.df}{data.frame of tweet messages and users}
}
\value{
link.df data.frame of links between users
}
\description{
Function to get user interaction links from tweets
}
