% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserLocation}
\alias{getUserLocation}
\title{Function to get tweets as clean data.frame from a tweets object}
\usage{
getUserLocation(users)
}
\arguments{
\item{user}{username of the twitter user}
}
\value{
geocode location
}
\description{
Function to get tweets as clean data.frame from a tweets object
}
