% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserTweets}
\alias{getUserTweets}
\title{Function to get tweets for a given user twitter handle or userInfo object}
\usage{
getUserTweets(username, n = 5)
}
\arguments{
\item{username}{twitter handle or user object of user whose followers need to be extracted.}
}
\value{
list of last n tweets by user
}
\description{
Function to get tweets for a given user twitter handle or userInfo object
}
