% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{getUserTweetsDataFrame}
\alias{getUserTweetsDataFrame}
\title{Function to get tweets as clean data.frame for a list of users}
\usage{
getUserTweetsDataFrame(usernames)
}
\arguments{
\item{usernames}{vector of one or more usernames}
}
\value{
tweets.df a data.frame with tweets details
}
\description{
Function to get tweets as clean data.frame for a list of users
}
