% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{setupAuthentication}
\alias{setupAuthentication}
\title{Function to setup authentication}
\usage{
setupAuthentication(filepath)
}
\arguments{
\item{path}{of file with authentication details}
}
\description{
Function to setup authentication
}

