% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicAPI.R
\name{whenLastActive}
\alias{whenLastActive}
\title{Function to get when was user last active}
\usage{
whenLastActive(userInfo)
}
\arguments{
\item{userInfo}{user object of the user}
}
\value{
time when user was last active
}
\description{
Function to get when was user last active
}

